undo scenario file relocate as it pollutes the PR too much
