+++ src/OpenLoco/CompanyManager.cpp
@@ -18,8 +18,8 @@
 #include "Objects/RoadObject.h"
 #include "Objects/TrackObject.h"
 #include "OpenLoco.h"
-#include "Scenario/Scenario.h"
-#include "Scenario/ScenarioManager.h"
+#include "Scenario.h"
+#include "ScenarioManager.h"
 #include "TownManager.h"
 #include "Ui/WindowManager.h"
 #include "Vehicles/Vehicle.h"
