+++ src/OpenLoco/OpenLoco.cpp
@@ -55,8 +55,8 @@
 #include "Platform/Crash.h"
 #include "Platform/Platform.h"
 #include "S5/S5.h"
-#include "Scenario/Scenario.h"
-#include "Scenario/ScenarioManager.h"
+#include "Scenario.h"
+#include "ScenarioManager.h"
 #include "StationManager.h"
 #include "Title.h"
 #include "TownManager.h"
