+++ src/OpenLoco/Scenario.cpp
@@ -1,36 +1,36 @@
 #include "./Scenario.h"
-#include "../Audio/Audio.h"
-#include "../CompanyManager.h"
-#include "../Date.h"
-#include "../Economy/Economy.h"
-#include "../Entities/EntityManager.h"
-#include "../Environment.h"
-#include "../Game.h"
-#include "../GameException.hpp"
-#include "../GameState.h"
-#include "../Graphics/Gfx.h"
-#include "../Gui.h"
-#include "../IndustryManager.h"
-#include "../Interop/Interop.hpp"
-#include "../Localisation/StringIds.h"
-#include "../Map/AnimationManager.h"
-#include "../Map/MapGenerator.h"
-#include "../Map/TileManager.h"
-#include "../Map/WaveManager.h"
-#include "../MultiPlayer.h"
-#include "../Objects/CargoObject.h"
-#include "../Objects/ClimateObject.h"
-#include "../Objects/ObjectManager.h"
-#include "../Objects/ScenarioTextObject.h"
-#include "../Platform/Platform.h"
-#include "../S5/S5.h"
-#include "../StationManager.h"
-#include "../Title.h"
-#include "../TownManager.h"
-#include "../Ui/WindowManager.h"
-#include "../Windows/Construction/Construction.h"
+#include "./Audio/Audio.h"
+#include "./CompanyManager.h"
+#include "./Date.h"
+#include "./Economy/Economy.h"
+#include "./Entities/EntityManager.h"
+#include "./Environment.h"
+#include "./Game.h"
+#include "./GameException.hpp"
+#include "./GameState.h"
+#include "./Graphics/Gfx.h"
+#include "./Gui.h"
+#include "./IndustryManager.h"
+#include "./Interop/Interop.hpp"
+#include "./Localisation/StringIds.h"
+#include "./Map/AnimationManager.h"
+#include "./Map/MapGenerator.h"
+#include "./Map/TileManager.h"
+#include "./Map/WaveManager.h"
+#include "./MultiPlayer.h"
+#include "./Objects/CargoObject.h"
+#include "./Objects/ClimateObject.h"
+#include "./Objects/ObjectManager.h"
+#include "./Objects/ScenarioTextObject.h"
+#include "./Platform/Platform.h"
+#include "./S5/S5.h"
 #include "./ScenarioManager.h"
 #include "./ScenarioObjective.h"
+#include "./StationManager.h"
+#include "./Title.h"
+#include "./TownManager.h"
+#include "./Ui/WindowManager.h"
+#include "./Windows/Construction/Construction.h"
 
 using namespace OpenLoco::Interop;
 using namespace OpenLoco::Map;
