+++ src/OpenLoco/ScenarioManager.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../Objects/Object.h"
+#include "./Objects/Object.h"
 #include "./Scenario.h"
 
 #include <cstddef>
