+++ src/OpenLoco/ScenarioObjective.cpp
@@ -1,5 +1,5 @@
 #include "ScenarioObjective.h"
-#include "../GameState.h"
+#include "./GameState.h"
 
 namespace OpenLoco::Scenario
 {
