+++ src/OpenLoco/ScenarioObjective.h
