+++ src/OpenLoco/Tutorial.cpp
@@ -5,7 +5,7 @@
 #include "Gui.h"
 #include "Interop/Interop.hpp"
 #include "Localisation/StringIds.h"
-#include "Scenario/Scenario.h"
+#include "Scenario.h"
 #include "Ui.h"
 
 #include <fstream>
