+++ src/OpenLoco/Windows/CompanyWindow.cpp
@@ -18,8 +18,8 @@
 #include "../Objects/InterfaceSkinObject.h"
 #include "../Objects/ObjectManager.h"
 #include "../OpenLoco.h"
-#include "../Scenario/Scenario.h"
-#include "../Scenario/ScenarioManager.h"
+#include "../Scenario.h"
+#include "../ScenarioManager.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
