+++ src/OpenLoco/Windows/ScenarioOptions.cpp
@@ -10,9 +10,9 @@
 #include "../Objects/ObjectManager.h"
 #include "../Objects/ScenarioTextObject.h"
 #include "../S5/S5.h"
-#include "../Scenario/Scenario.h"
-#include "../Scenario/ScenarioManager.h"
-#include "../Scenario/ScenarioObjective.h"
+#include "../Scenario.h"
+#include "../ScenarioManager.h"
+#include "../ScenarioObjective.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/WindowManager.h"
 #include "../Widget.h"
