+++ src/OpenLoco/Windows/ScenarioSelect.cpp
@@ -8,8 +8,8 @@
 #include "../Localisation/FormatArguments.hpp"
 #include "../Localisation/StringIds.h"
 #include "../Objects/ObjectManager.h"
-#include "../Scenario/Scenario.h"
-#include "../Scenario/ScenarioManager.h"
+#include "../Scenario.h"
+#include "../ScenarioManager.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
 #include "../Widget.h"
