+++ src/OpenLoco/Windows/TerraForm.cpp
@@ -19,7 +19,7 @@
 #include "../Objects/TreeObject.h"
 #include "../Objects/WallObject.h"
 #include "../Objects/WaterObject.h"
-#include "../Scenario/Scenario.h"
+#include "../Scenario.h"
 #include "../Ui/Dropdown.h"
 #include "../Ui/ScrollView.h"
 #include "../Ui/WindowManager.h"
